+++ mmdet/models/backbones/__init__.py
@@ -1,6 +1,5 @@
 from .hrnet import HRNet
-from .resnet import ResNet
-from .resnet_v1d import ResNetV1d
+from .resnet import ResNet, ResNetV1d
 from .resnext import ResNeXt
 from .ssd_vgg import SSDVGG
